fixed warnings
git-svn-id: svn://svn.icculus.org/gtkradiant/GtkRadiant/trunk@79 8a3a26a2-13c4-0310-b231-cf6edde360e5

+++ contrib/bobtoolz/DBrush.cpp
@@ -126,7 +126,7 @@ int DBrush::BuildPoints()
 //	Sys_Printf("%i points on brush\n", pointList.size());
 #endif
 
-	return pointList.size();
+	return static_cast<int>(pointList.size());
 }
 
 void DBrush_addFace(DBrush& brush, const _QERFaceData& faceData)
